fix: Switchコンポーネントのファイル階層を移動
