+++ src/stories/atoms/Switch.module.css
@@ -1,0 +1,60 @@
+/* ファイル: src/components/atoms/Switch/Switch.module.css */
+
+.label {
+  position: relative;
+  display: inline-block;
+  width: 50px;
+  height: 24px;
+}
+
+.input {
+  /* ネイティブ checkbox を完全に隠す */
+  opacity: 0;
+  width: 0;
+  height: 0;
+  position: absolute;
+}
+
+.track {
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background-color: #ccc;
+  border-radius: 24px;
+  transition: background-color 0.2s;
+}
+
+.thumb {
+  position: absolute;
+  top: 3px;
+  left: 3px;
+  width: 18px;
+  height: 18px;
+  background-color: #fff;
+  border-radius: 50%;
+  transition: transform 0.2s;
+}
+
+/* フォーカス時の視覚的なインジケーター */
+.input:focus + .track {
+  box-shadow: 0 0 0 3px rgba(33, 150, 243, 0.5);
+}
+
+/* チェック時の軌道色とノブの位置変化 */
+.input:checked + .track {
+  background-color: #2196f3;
+}
+.input:checked + .track .thumb {
+  transform: translateX(26px);
+}
+
+/* disabled 時の見た目 */
+.input:disabled + .track {
+  background-color: #eee;
+  cursor: not-allowed;
+}
+.input:disabled + .track .thumb {
+  background-color: #ddd;
+}
