--- src/stories/atoms/TextArea.stories.ts
@@ -1,23 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react-vite';
-import { TextArea } from './TextArea';
-
-const meta = {
-  title: 'Atoms/Textarea',
-  component: TextArea,
-  parameters: {
-    layout: 'centered',
-  },
-  tags: ['autodocs'],
-  args: {},
-} satisfies Meta<typeof TextArea>;
-
-export default meta;
-type Story = StoryObj<typeof meta>;
-
-export const Primary: Story = {
-  args: {
-    label: 'テキストエリア',
-    value: 'Valueサンプル',
-    id: 'textarea-example',
-  },
-};
