--- src/stories/atoms/dropdown.css
@@ -1,58 +1,0 @@
-.dropdown {
-  display: flex;
-  width: 475px;
-  flex-direction: column;
-  align-items: flex-start;
-  gap: 5px;
-}
-
-.dropdown label {
-  display: flex;
-  align-items: flex-start;
-  gap: 8px;
-  align-self: stretch;
-  color: #000;
-  font-size: 15px;
-  font-style: normal;
-  font-weight: 700;
-  line-height: normal;
-}
-
-.dropdown label span {
-  color: #ff544b;
-  font-family: Inter;
-  font-size: 15px;
-  font-style: normal;
-  font-weight: 700;
-  line-height: normal;
-}
-
-.dropdownSelect {
-  display: flex;
-  padding: 13px 16px;
-  align-items: center;
-  align-self: stretch;
-  border-radius: 4px;
-  border: 1px solid #bcc0c8;
-  background: #fff;
-}
-
-.dropdownSelect option {
-  color: #292929;
-  font-family: Inter;
-  font-size: 14px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: normal;
-}
-
-.warningText {
-  height: 19px;
-  align-self: stretch;
-  color: #ff544b;
-  font-family: Inter;
-  font-size: 14px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: normal;
-}
